requirements.txt
appnope==0.1.0
asgiref==3.2.7
astroid==2.4.2
attrs==19.3.0
autopep8==1.5.3
backcall==0.2.0
certifi==2020.6.20
cloudinary==1.21.1
coverage==5.2.1
decorator==4.4.2
dj-database-url==0.5.0
Django==3.0.7
django-appconf==1.0.4
django-extensions==3.0.4
django-filter==2.3.0
django-measurement==3.2.3
djangorestframework==3.11.1
flake8==3.8.3
gunicorn==20.0.4
iniconfig==1.0.1
ipdb==0.13.3
ipython==7.17.0
ipython-genutils==0.2.0
isort==4.3.21
jedi==0.17.2
lazy-object-proxy==1.4.3
mccabe==0.6.1
measurement==3.2.0
more-itertools==8.4.0
mpmath==1.1.0
packaging==20.4
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.1.2
pluggy==0.13.1
prompt-toolkit==3.0.5
psycopg2-binary==2.8.5
ptyprocess==0.6.0
py==1.9.0
pycodestyle==2.6.0
pyflakes==2.2.0
Pygments==2.6.1
pylint==2.5.3
pyparsing==2.4.7
pytest==6.0.1
pytest-cov==2.10.0
pytest-django==3.9.0
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
sympy==1.6
toml==0.10.1
traitlets==4.3.3
typed-ast==1.4.1
urllib3==1.25.9
wcwidth==0.2.5
whitenoise==5.1.0
wrapt==1.12.1
